moved auth functions to seperate auth file
