+++ back-end/server/api/authentication.py
@@ -1,1 +1,27 @@
 # need to implement CSRF
+from rest_framework.authentication import CSRFCheck
+from rest_framework_simplejwt.authentication import JWTAuthentication
+from rest_framework import exceptions
+from server.settings import SIMPLE_JWT
+
+def enforce_csrf(request):
+  check = CSRFCheck()
+  check.process_request(request)
+  reason = check.process_view(request, None, (), {})
+  if reason:
+    raise exceptions.PermissionDenied('CSRF failed: %s' % reason)
+
+class CustomAuthentication(JWTAuthentication):
+  def authenticate(self, request):
+    header = self.get_header(request)
+
+    if header is None:
+      raw_token = request.COOKIES.get(SIMPLE_JWT['AUTH_COOKIE']) or None
+    else:
+      raw_token = self.get_raw_token(header)
+    if raw_token is None:
+      return None
+
+    validated_token = self.get_validated_token(raw_token)
+    enforce_csrf(request)
+    return self.get_user(validated_token), validated_token
