+++ back-end/server/api/views.py
@@ -4,11 +4,8 @@ from rest_framework.views import APIView
 from django.middleware import csrf
 from django.contrib.auth import authenticate
 from rest_framework_simplejwt.tokens import RefreshToken
-from rest_framework.authentication import CSRFCheck
-from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework_simplejwt.views import TokenRefreshView
 from rest_framework.decorators import api_view
-from rest_framework import exceptions
 from server.settings import SIMPLE_JWT
 
 # Create your views here.
@@ -27,28 +24,6 @@ def get_tokens_for_user(user):
         'refresh': str(refresh),
         'access': str(refresh.access_token),
     }
-
-def enforce_csrf(request):
-  check = CSRFCheck()
-  check.process_request(request)
-  reason = check.process_view(request, None, (), {})
-  if reason:
-    raise exceptions.PermissionDenied('CSRF failed: %s' % reason)
-
-class CustomAuthentication(JWTAuthentication):
-  def authenticate(self, request):
-    header = self.get_header(request)
-
-    if header is None:
-      raw_token = request.COOKIES.get(SIMPLE_JWT['AUTH_COOKIE']) or None
-    else:
-      raw_token = self.get_raw_token(header)
-    if raw_token is None:
-      return None
-
-    validated_token = self.get_validated_token(raw_token)
-    enforce_csrf(request)
-    return self.get_user(validated_token), validated_token
 
 class CustomTokenView(TokenRefreshView):
   def post(self, request, *args, **kwargs):
